added a possible approach to separating the handling of raw, untransformed, data from article xml with rawJATS.py and shifting the presentation into parseJATS.py. only seven or eight functions have been done thus far as an example, but all of the lettuce tests pass as well as my own quick and very dirty unnittests
